Update file names on preconditioners
